change uint -> uint256

+++ contracts/core/Vesting.sol
@@ -19,12 +19,12 @@ contract Vesting is Ownable {
     // Token percent to be minted in specific quarter
     mapping (uint256 => uint256) public tokensUnlocking;
     // Check the tokens of specific quarter is minted.
-    mapping (address => mapping(uint => bool)) public isMinted;
+    mapping (address => mapping(uint256 => bool)) public isMinted;
 
-    event Minted(address indexed _user, uint indexed _quarter, uint _amount);
+    event Minted(address indexed _user, uint256 indexed _quarter, uint256 _amount);
     event RemovedBasisPoint(address _user);
     event AssignedBasisPoint(address _user, uint32 _basePoint);
-    event UpdatedTokensUnlocking(uint _quarter, uint _tokensUnlocking);
+    event UpdatedTokensUnlocking(uint256 _quarter, uint256 _tokensUnlocking);
 
     /** 
         @notice Contructor
@@ -33,22 +33,22 @@ contract Vesting is Ownable {
         @param  _quarters  uint256[] Quarter timestamp array
         @param  _tokensUnlocking  uint32[] Tokens to be unlocked
      */
-    constructor(address _mariposa, address[] memory _ownerships, uint32[] memory _basisPoints, uint[] memory _quarters, uint[] memory _tokensUnlocking) {
+    constructor(address _mariposa, address[] memory _ownerships, uint32[] memory _basisPoints, uint256[] memory _quarters, uint256[] memory _tokensUnlocking) {
         mariposa = IMariposa(_mariposa);
-        uint ownershipsLen = _ownerships.length;
+        uint256 ownershipsLen = _ownerships.length;
         require(ownershipsLen == _basisPoints.length, "Vesting: length error");
 
-        uint quartersLen = _quarters.length;
+        uint256 quartersLen = _quarters.length;
         require(quartersLen == _tokensUnlocking.length, "Vesting: length error");
         uint32 basePointsSum;
-        for (uint i; i < ownershipsLen; i ++) {
+        for (uint256 i; i < ownershipsLen; i ++) {
             basisPoints[_ownerships[i]] = _basisPoints[i];
             basePointsSum += _basisPoints[i];
         }
         
         require(basePointsSum == 1e8, "Vesting: checksum error");
 
-        for (uint i; i < quartersLen;i ++)
+        for (uint256 i; i < quartersLen;i ++)
             tokensUnlocking[_quarters[i]] = _tokensUnlocking[i];
     }
 
@@ -56,12 +56,12 @@ contract Vesting is Ownable {
         @notice Mints tokens to recipient 
         @param  _quarter  uint256 Quarter timestamp to mint
      */
-    function mint(uint _quarter) 
+    function mint(uint256 _quarter) 
         external
     {
         require(_quarter > 0 && _quarter < block.timestamp, "Vesting: can not mint");
         require(isMinted[msg.sender][_quarter] == false, "Vesting: already minted");
-        uint mintAmount = tokensUnlocking[_quarter] * basisPoints[msg.sender];
+        uint256 mintAmount = tokensUnlocking[_quarter] * basisPoints[msg.sender];
         isMinted[msg.sender][_quarter] = true;
         mariposa.request(msg.sender, mintAmount);
         emit Minted(msg.sender, _quarter, mintAmount);
@@ -96,7 +96,7 @@ contract Vesting is Ownable {
         @param  _quarter  uint256 Quarter timestamp 
         @param  _tokensUnlocking  uint256 Tokens to be unlocked
      */
-    function updateTokensUnlocking(uint _quarter, uint _tokensUnlocking) external onlyOwner {
+    function updateTokensUnlocking(uint256 _quarter, uint256 _tokensUnlocking) external onlyOwner {
         require(_quarter > 0, "Vesting: zero quarter");
         tokensUnlocking[_quarter] = _tokensUnlocking;
         emit UpdatedTokensUnlocking(_quarter, _tokensUnlocking);
